+++ docs/source/conf.py
@@ -263,9 +263,6 @@ intersphinx_mapping = {
     "numpy": ("https://numpy.org/doc/stable/", None),
     "scipy": ("https://docs.scipy.org/doc/scipy/", None),
     "matplotlib": ("https://matplotlib.org/stable/", None),
-    "jax": ("https://jax.readthedocs.io/en/latest/", None),
-    "flax": ("https://flax.readthedocs.io/en/latest/", None),
-    "ray": ("https://docs.ray.io/en/latest/", None),
 }
 # Added timeout due to periodic scipy.org down time
 # intersphinx_timeout = 30
