+++ scripts/reconstruct_demo.py
@@ -12,107 +12,62 @@ import numpy as np
 
 import matplotlib.pyplot as plt
 
-from trinidi import cross_section, reconstruct, resolution, util
+from trinidi import cross_section, reconstruct, resolution, simulate, util
 
 """
 Generation of Simulated data
 ----------------------------
-We use the function below to generate the data. The details of this function are not necessarily
-important for the user to understand.
-"""
+To generate the simulated data we use `trinidi`'s `generate_sample_data` function that generates
+a phantom of overlapping discs of different isotopes and the associated measurement counts.
 
-
-def generate_sample_data(
-    isotopes,
-    z,
-    Δt=0.90,
-    t_0=72,
-    t_last=400,
-    flight_path_length=10,
-    projection_shape=(31, 31),
-    kernels=None,
-    acquisition_time=10,
-):
-    r"""Generate example data."""
-
-    t_A = np.arange(t_0, t_last, Δt)
-    N_A = t_A.size
-
-    if not kernels:
-        kernels = [np.array([1]), np.array([1 / 4, 1 / 4, 1 / 4, 1 / 4])]
-
-    output_shape = projection_shape + (N_A,)
-
-    R = resolution.ResolutionOperator(output_shape, t_A, kernels=kernels)
-    t_F = R.t_F
-
-    ϕ, b, θ, α_1, α_2 = util.generate_spectra(t_A, acquisition_time=10)
-    N_b = θ.size
-
-    D = cross_section.XSDict(isotopes, t_F, flight_path_length)
-
-    Z = util.rose_phantom(projection_shape[0], num_circles=z.size, radius=2 / 3) * z.reshape(
-        [1, 1, z.size]
-    )
-
-    v = np.random.poisson(1000, size=projection_shape + (1,))
-    v = v / v.mean()
-
-    Φ = v @ ϕ.T
-    B = v @ b.T
-
-    Y_o_bar = Φ + B
-    Y_s_bar = α_1 * (Φ * R(np.exp(-Z @ D.values)) + α_2 * B)
-
-    Y_o = np.random.poisson(Y_o_bar)
-    Y_s = np.random.poisson(Y_s_bar)
-
-    ground_truth_params = {"z": z, "α_1": α_1, "α_2": α_2, "θ": θ}
-
-    return Y_o, Y_s, ground_truth_params, Z, t_A, flight_path_length, N_b, kernels
-
-
-"""
-Below we define what isotopes to use for the phantom and what their corresponding densities should
-become.
+Below we define what isotopes to use for the phantom and what their corresponding densities, `z`
+should become. The `projection_shape` is 2D and is the the size of the detector, but in general it
+can be of any dimensionality.
 """
 
 isotopes = ["U-238", "Pu-239", "Ta-181"]
 z = np.array([[0.005, 0.003, 0.004]]).T
 
 
-# projection_shape = (20, 20) # use this for fast execution
-projection_shape = (60, 60)  # use this for the docs
+# projection_shape = (16, 16) # use this for fast execution
+projection_shape = (64, 64)  # use this for the docs
 
-Y_o, Y_s, ground_truth_params, Z, t_A, flight_path_length, N_b, kernels = generate_sample_data(
+(
+    Y_o,
+    Y_s,
+    ground_truth_params,
+    Z,
+    t_A,
+    flight_path_length,
+    N_b,
+    kernels,
+) = simulate.generate_sample_data(
     isotopes, z, acquisition_time=10, projection_shape=projection_shape
 )
 
 """
-The arrays `Y_o` and `Y_s` contain the the open beam and sample neutron count measurements.
-The first (two) axes correspond to `projection_shape`, which in most cases is equal to the
-detector shape. The last axis correponds to the time-of-arrival (TOA) dimension which has
-`N_A` bins.
+The arrays `Y_o` and `Y_s` contain the the open beam and sample neutron count measurements. The
+first (two) axes correspond to `projection_shape`, which in most cases is equal to the detector
+shape. The last axis correponds to the time-of-arrival (TOA) dimension which has `N_A` bins.
 """
+N_A = Y_o.shape[-1]
+
 print(f"{Y_o.shape = }     (Open beam measurement)")
 print(f"{Y_s.shape = }     (Sample measurement)")
-print()
-projection_shape = Y_o.shape[:-1]
 print(f"{projection_shape = }     (Shape of the detector)")
-print()
-N_A = Y_o.shape[-1]
 print(f"{N_A = }     (Number of TOA bins)")
 
 
 """
-The corresponding array of TOAs (`t_A`) is in units of [μs] and has the number of elements, `N_A`.
+The corresponding array of TOAs (`t_A`) is in units of [μs] and has `N_A` elements.
 """
 print(f"t_A = [{t_A[0]:.2f}, {t_A[1]:.2f}, ..., {t_A[-2]:.2f}, {t_A[-1]:.2f}] [μs]")
-print(f"{t_A.shape = }")
+print(f"{t_A.shape = }     (Same as N_A)")
 
 
 """
-The `flight_path_length` has units of [m] a relates the neutron times with the neutron energies.
+The `flight_path_length` has units of [m] and relates the neutron times with the neutron energies.
+Below we illustrate the conversion from time to energy.
 """
 print(f"{flight_path_length = } [m]")
 E = util.time2energy(t_A, flight_path_length)
@@ -120,24 +75,27 @@ print(f"E = [{E[0]:.2f}, {E[1]:.2f}, ..., {E[-2]:.2f}, {E[-1]:.2f}] [eV]")
 
 
 """
-The ground truth areal densities `Z` are used to compare to our estimates. `Z` has the same
-`projection_shape` as the measurements. The last axis of `Z` is the number of isotopes.
+The ground truth areal densities `Z` are used in order to later compare to our estimates. `Z` has
+the same `projection_shape` as the measurements. The last axis of `Z` is the number of isotopes,
+`N_m`.
 """
-print(f"{Z.shape = }     (Ground truth areal densities)")
-
 N_m = Z.shape[-1]
+
+print(f"{Z.shape = }     (Ground truth areal densities)")
 print(f"{N_m = }     (Number of isotopes)")
-print(f"{len(isotopes) = }")
+print(f"{len(isotopes) = }     (Also number of isotopes))")
 
 
 """
-We can show the areal densities `Z` using the plotting function below. The phantom consists of
-several discs and disc `i` has the density `z[i]` and corresponds to `isotope[i]`.
+We can show the areal densities `Z` using the plotting function below.
 """
 
 
-def plot_densities(fig, ax, Z, isotopes, vmaxs=None):
+def plot_densities(Z, isotopes, vmaxs=None):
     r"""Show areal densities. `ax` must be list."""
+
+    fig, ax = plt.subplots(1, len(isotopes), figsize=[12, 3.3])
+    ax = np.atleast_1d(ax)
 
     for i, isotope in enumerate(isotopes):
         z = Z[:, :, i]
@@ -156,44 +114,50 @@ def plot_densities(fig, ax, Z, isotopes, vmaxs=None):
     return fig, ax
 
 
-fig, ax = plt.subplots(1, N_m, figsize=[12, 3.3])
-ax = np.atleast_1d(ax)
-plot_densities(fig, ax, Z, isotopes, vmaxs=z * 1.5)
+plot_densities(Z, isotopes, vmaxs=z * 1.5)
 plt.show()
 
+"""
+As shown above, the phantom consists of several discs and the `i`th disc corresponds to
+`isotopes[i]` which has density `z[i]`, so it is equal to `Z[:,:,i].max()`.
 
-print(f"{z.T = } [mol/cm²]")
-print(f"{isotopes = }")
+"""
 
+for i, (iso, z_i) in enumerate(zip(isotopes, z.flatten())):
+    print(f"{iso}: {z_i} = ({Z[:,:,i].max()}) [mol/cm²]")
 
 """
 Preparation for the Reconstruction
 ----------------------------------
 
-For the reconstruction we need to define the regions `Ω_z` and `Ω_0` that correspond to the uniformly
-dense region and the open beam region, respectively.
+For the reconstruction we need to define the regions `Ω_z` and `Ω_0` that correspond to the so
+called uniformly dense region and the open beam region, respectively.
 
-For this we use the `ProjectionRegion` class. We initialize it with a mask (boolean array) that is
-of shape `projection_shape + (1,)`, indicating which pixels belong to these regions.
+For this we use the `ProjectionRegion` class. We initialize each with a mask (boolean array) that
+has shape `projection_shape + (1,)`, indicating which pixels belong to these regions.
 
 In this example we use the ground truth `Z` to find these regions as
- - the overlap of all discs (Ω_z)
- - the complement of the union of all discs (Ω_0).
+ - the overlap of all discs: Ω_z
+ - the complement of the union of all discs: Ω_0.
 
-(In the case when the ground truth is not known, the user will have to find a way to define these
-regions.)
+(In the general case when the ground truth is not known, the user will have to find a way to define
+these regions such as through prior knowledge about where to adequately find such regions.)
+"""
+mask_z = np.prod(Z, axis=2, keepdims=True) > 0
+Ω_z = reconstruct.ProjectionRegion(mask_z)
+
+mask_0 = np.sum(Z, axis=2, keepdims=True) == 0
+Ω_0 = reconstruct.ProjectionRegion(mask_0)
+
+print(f"mask_z.shape = mask_0.shape = {mask_0.shape}")
+
+
+"""
+We can illustrate these regions using the `ProjectionRegion.plot_contour` and the
+`ProjectionRegion.imshow` functions below.
 """
 
-Ω_z = reconstruct.ProjectionRegion(np.prod(Z, axis=2, keepdims=True) > 0)
-Ω_0 = reconstruct.ProjectionRegion(np.sum(Z, axis=2, keepdims=True) == 0)
-
-
-"""
-We can illustrate these regions using the builtin `ProjectionRegion.plot_contour`
-function and the `ProjectionRegion.imshow` function below.
-"""
-
-fig, ax = plt.subplots(1, 3, figsize=[14, 4])
+fig, ax = plt.subplots(1, 3, figsize=[12, 4])
 ax = np.atleast_1d(ax)
 
 ax[0].imshow(np.sum(Y_s, axis=-1) / np.sum(Y_o, axis=-1), vmin=0)
@@ -216,8 +180,8 @@ print(R)
 
 """
 Since the resolution operator inputs wider spectra than it outputs, there is an additional vector
-of time-of-flights (TOF), `t_F`, with size `N_F`. The `t_F` array is calibrated such that
-approximately R(t_A) = t_F.
+of time-of-flights (TOF), `t_F`, with size `N_F`, associated with the input times. The `t_F` array
+is calibrated such that approximately R(t_A) = t_F.
 """
 
 t_F = R.t_F
@@ -240,9 +204,6 @@ print(D)
 """
 Below we plot the average measurements in the `Ω_z` and `Ω_0` regions compared to the cross section
 dictionary. We use the `ProjectionRegion.averge` function to compute these average spectra.
-
-(Note that in `trinidi` we define vecors with `N` elements to have shape `(N,1)` and thus we
-occasionally have to explicitly flatten the arrays for correct plotting etc.)
 """
 
 fig, ax = plt.subplots(2, 1, figsize=[12, 8], sharex=True)
@@ -258,28 +219,31 @@ plt.show()
 """
 Reconstruction of the Nuisance Parameters
 -----------------------------------------
+The nuisance parameters are computed first before the densities can be computed. They parametrize
+such quantities associated the the neutron flux and background at every pixel (Φ, B) and global
+scalers such as the overall relative exposure and sample induced background factors (α_1, α_2). We
+do not expect the general user to need to understand these parameters intricately and thus the
+handling of these parameters is encapsulated inside the `ParameterEstimator` object.
 """
 
 
 """
-We define a `ParameterEstimator` object. For the background we use the same number `N_b` of basis
-functions. This estimates the nuisance parameters (z, α_1, α_1, θ). Novice users do not need to
-worry about the values and handling of these parameters.
+Below we define a `ParameterEstimator` object. For the background we use the same number `N_b` of
+basis functions as used for the simulation. This estimates the nuisance parameters
+(z, α_1, α_1, θ) using the relatively fast `BFGS` algorithm.
 """
 par = reconstruct.ParameterEstimator(Y_o, Y_s, R, D, Ω_z, Ω_0=Ω_0, N_b=N_b)
 
 
 """
-`Advanced users`: The parameters are available through the `ParameterEstimator.get` function.
+`Advanced users:` The parameters are available through the `ParameterEstimator.get` function. They
+can manually be modified using the `ParameterEstimator.set` function. We also provide similar
+`save` and `load` functions to write/read them to file.
 """
 d = par.get()
 print(d)
 
 
-"""
-`Advanced users`: The parameters can manually be modified using the `ParameterEstimator.set` function.
-We also provide similar `save` and `load` functions to write/read them to file.
-"""
 par.set(**d)
 par.set(z=d["z"], α_1=d["α_1"], α_2=d["α_2"], θ=d["θ"])  # same as line above
 
@@ -287,53 +251,36 @@ par.set(z=d["z"], α_1=d["α_1"], α_2=d["α_2"], θ=d["θ"])  # same as line ab
 # d = par.load("par.npy")
 
 """
-Similar to above we can easily plot the `Ω_z` and `Ω_0` regions.
+Once the `ParameterEstimator` object has been constructed we can plot the `Ω_z` and `Ω_0` regions
+much more easily than above using the `ParameterEstimator.plot_regions` function.
 """
 par.plot_regions()
 plt.show()
 
 """
 The `ParameterEstimator.plot_results` function allows to display the resulting spectra from the
-estimation.
+estimation and verify good fitting of the signals.
 """
 par.plot_results()
 plt.show()
 
 
 """
-We also provide the APGM optimization routine, however we do not recommend it since tends to be
-signicicantly slower.
+We also provide the APGM optimization routine used in our paper, however we do not recommend it
+since tends to be signicicantly slower.
 """
 
-# if False:
-#     par.apgm_solve(iterations=100)
-#     fig, ax = par.apgm_plot_convergence(plot_residual=True, ground_truth=ground_truth_params)
-#     plt.show()
-
-# par.set(**d)
+# par.apgm_solve(iterations=100)
+# fig, ax = par.apgm_plot_convergence(plot_residual=True, ground_truth=ground_truth_params)
+# plt.show()
 
 
 """
 Reconstruction of the Areal Densities
 -------------------------------------
+Now that the nuisance parameters have been computed and stored in the `ParameterEstimator` object,
+we can carry on with the main task, i.e. with the areal density reconstruction.
 """
-
-"""
-We use this simple plotting function below to keep things clean.
-"""
-
-
-def plot_compare(Z, str_Z, Z_hat, str_Z_hat):
-    r"""Generate two plots comparing ground truth with reconstruction."""
-    fig, ax = plt.subplots(1, N_m, figsize=[12, 3.3])
-    ax = np.atleast_1d(ax)
-    plot_densities(fig, ax, Z, isotopes, vmaxs=z * 1.5)
-    fig.suptitle(f"{str_Z} [mol/cm²]")
-
-    fig, ax = plt.subplots(1, N_m, figsize=[12, 3.3])
-    ax = np.atleast_1d(ax)
-    plot_densities(fig, ax, Z_hat, isotopes, vmaxs=z * 1.5)
-    fig.suptitle(f"{str_Z_hat} [mol/cm²]")
 
 
 """
@@ -347,7 +294,8 @@ constructor.)
 den = reconstruct.DensityEstimator(Y_s, par, non_negative_Z=False, dispperiod=50)
 
 """
-We reonstruct the areal densities using the `DensityEstimator.solve` function.
+We reonstruct the areal densities using the `DensityEstimator.solve` function which returns the
+areal density estimates `Z_hat`.
 """
 Z_hat = den.solve(iterations=200)
 
@@ -362,8 +310,19 @@ den.plot_convergence(ground_truth=Z)
 plt.show()
 
 """
-We plot the resulting areal density estimates below and compare them to the ground truth.
+We use this plotting function below to plot the resulting areal density estimates and compare them
+to the ground truth (`Z`)
 """
+
+
+def plot_compare(Z, str_Z, Z_hat, str_Z_hat):
+    r"""Generate two plots comparing ground truth with reconstruction."""
+    fig, ax = plot_densities(Z, isotopes, vmaxs=z * 1.5)
+    fig.suptitle(f"{str_Z} [mol/cm²]")
+
+    fig, ax = plot_densities(Z_hat, isotopes, vmaxs=z * 1.5)
+    fig.suptitle(f"{str_Z_hat} [mol/cm²]")
+
 
 plot_compare(Z, "Ground Truth", Z_hat, "Reconstruction")
 plt.show()
@@ -373,10 +332,11 @@ plt.show()
 Cropped or Binned Measurements
 ------------------------------
 
-In practice, it is not always desired to reconstruct the full field of view or at full resolution.
-For this reason we provide the `projection_transform` argument in the `DensityEstimator`
-constructor. Perhaps we require the full field of view to estimate the nuisance parameters, but
+In practice, it is not always desired to reconstruct the full field of view or at full resolution:
+Perhaps we require the full field of view to estimate the nuisance parameters, but
 we only want to reconstruct a small cropped region.
+For this reason we can provide the `projection_transform` argument in the `DensityEstimator`
+constructor.
 
 The `projection_transform` is the desired operation that is applied to the measurements and used
 internally to handle the modification of the background and flux estimates.
@@ -424,7 +384,7 @@ plt.show()
 
 
 """
-Second we show an example of cropping data using the below defined `crop` function and the same
+Second we show an example of cropping data using the below defined `crop` function.
 """
 
 
@@ -450,3 +410,8 @@ print(f"{Z.shape = }")
 
 plot_compare(Z_crop, "Cropped Ground Truth", Z_hat_crop, "Reconstruction from Cropped Measurement")
 plt.show()
+
+"""
+If multiple transformations are required, those can easily be composed such as below.
+"""
+projection_transform = lambda Y: crop(binning_2x2(Y))
